intelligent_agent/requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM Providers
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-groq>=0.1.0
langchain-community>=0.3.0

# Search Tools
tavily-python>=0.5.0

# Frontend
streamlit>=1.32.0

# Configuration & Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
requests>=2.32.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Optional: For HuggingFace (if using)
transformers>=4.40.0
torch>=2.0.0

# Optional: For local Ollama (if using)
langchain-ollama>=0.1.0

# Database & Persistence
langgraph-checkpoint-postgres>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0